// typescript/user-interface/packages/ui-app/src/ts/components/analyst-ui/components/location/types.ts
export interface LocationMutations {
  locateEvent: (args: any) => Promise<void>;
  createEvent: (args: any) => Promise<void>;
  updateFeaturePredictions: (args: any) => Promise<void>;
  updateDetections: (args: any) => Promise<void>;
  deleteDetections: (args: any) => Promise<void>;
  changeSignalDetectionAssociations: (args: any) => Promise<void>;
}
// ...
export interface LocationReduxProps {
  // Passed in from golden-layout
  glContainer?: GoldenLayout.Container;
  // The currently-open processing interval time range
  currentTimeInterval: CommonTypes.TimeRange;
  // The currently-open event hypothesis IDs
  openEventId: string;
  // The currently-selected signal detection IDs
  selectedSdIds: string[];
  // used for additional time range values
  analysisMode: WorkflowTypes.AnalysisMode;
  /** The measurement mode */
  measurementMode: AnalystWorkspaceTypes.MeasurementMode;
  /** The signal detections selected to show fk */
  sdIdsToShowFk: string[];
  /** The selected location state */
  location: AnalystWorkspaceTypes.LocationSolutionState;

  setOpenEventId(
    event: LegacyEventTypes.Event | undefined,
    latestLocationSolutionSet: LegacyEventTypes.LocationSolutionSet | undefined,
    preferredLocationSolutionId: string | undefined
  ): void;
  setSelectedEventIds(ids: string[]): void;
  setSelectedSdIds(ids: string[]): void;
  setMeasurementModeEntries(entries: Record<string, boolean>): void;
  setSelectedLocationSolution(locationSolutionSetId: string, locationSolutionId: string): void;
  setSelectedPreferredLocationSolution(
    preferredLocationSolutionSetId: string,
    preferredLocationSolutionId: string
  ): void;
}

/**
 * Location Props
 */
export type LocationProps = LocationReduxProps & LocationMutations;